feat: make shadow node work, share the code between ios
