+++ package.json
@@ -205,7 +205,7 @@
     "use-latest-callback": "^0.2.1"
   },
   "codegenConfig": {
-    "name": "RNCTabViewSpec",
+    "name": "RNCTabView",
     "type": "components",
     "jsSrcsDir": "src",
     "android": {
